refactor(renderer-components): relocate resource selectors/dialogs from features/ to components/
features/resource (introduced in #16187) is only a partial extraction — it still
imports pages/library/{adapters,constants,types}, so it is not a self-contained
features/ domain (a features/ module must not import external content; external
code should reference the domain once and get everything). Until that fuller
extraction happens, move the whole unit (Agent/Assistant/Workspace selectors +
create/edit dialogs) to components/resource and keep adapters/constants/types in
pages/library. Pure directory move + import-alias rewrite; no behavior change.
