+++ src/renderer/components/resource/ResourceSelectorShell.tsx
