+++ src/renderer/components/resource/WorkspaceSelector.tsx
