+++ src/renderer/components/resource/__tests__/AssistantSelector.test.tsx
