+++ src/renderer/components/resource/__tests__/ResourceSelectorShell.test.tsx
