+++ src/renderer/components/resource/__tests__/WorkspaceSelector.test.tsx
