+++ src/renderer/components/resource/dialogs/ResourceEditDialogHost.tsx
