+++ src/renderer/components/resource/dialogs/__tests__/EditDialogs.test.tsx
