+++ src/renderer/components/resource/dialogs/components/CatalogPicker.tsx
