+++ src/renderer/components/resource/dialogs/components/McpServerAvatar.tsx
