+++ src/renderer/components/resource/dialogs/components/TagSelector.tsx
