+++ src/renderer/components/resource/dialogs/edit/AssistantEditDialog.tsx
