+++ src/renderer/components/resource/dialogs/edit/EditDialogShared.tsx
