+++ src/renderer/components/resource/dialogs/form/__tests__/assistant.test.ts
