+++ src/renderer/components/resource/dialogs/form/assistant.ts
