+++ src/renderer/components/resource/dialogs/form/assistantModelFilter.ts
