+++ src/renderer/components/resource/dialogs/index.ts
