+++ src/renderer/components/resource/index.ts
